refactored routing and component structure
