+++ frontend/package.json
@@ -15,7 +15,7 @@
     "minidenticons": "^4.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-router-dom": "^6.16.0",
+    "react-router-dom": "^6.22.3",
     "sass": "^1.71.1",
     "use-sound": "^4.0.1",
     "websocket": "^1.0.34"
