+++ frontend/src/components/App.jsx
@@ -1,55 +1,26 @@
 /**
  * Main application component.
+ * SOMETHING WRONG WITH THE WAY WEBSOCKET STATE IS HANDLED CLIENT-SIDE WITH THIS NEW STRUCTURE. 
+ * ~~INITIATING WEBSOCKET CLOSE REQUESTS TO SERVER BUT FAILING TO ACTUALLY SEND THEM LOL~~
+ * ONLY HAPPENS ON SECOND CONSECUTIVE ENTRANCE INTO LOBBY OF THE SAME NAME.
+ * INCREDIBLE STUFF BY REACT, WHAT A FUN TIME.
  * @module App
  */
 
-import React, { useEffect, useRef, useState, useMemo } from 'react';
-import '../styles/app.scss';
-import { minidenticon } from 'minidenticons';
+import React, { useState, useRef, useEffect } from "react";
+import { Routes, Route } from 'react-router-dom';
+import '../styles/global.scss';
 import Lobby from './Lobby.jsx';
-import useSound from 'use-sound';
-import Enter from '../sounds/wheep-wheep.mp3';
-import Click from '../sounds/mouse-click.mp3';
-import Denied from '../sounds/keycard-denial.mp3';
-import mutedSVG from '../images/muted.svg';
-import unmutedSVG from '../images/unmuted.svg';
-
-/**
- * Displays a minidenticon image. Credit: https://github.com/laurentpayot/minidenticons
- * @param {object} props Component props.
- * @param {string} props.username Username.
- * @param {string} props.saturation Saturation value.
- * @param {string} props.lightness Lightness value.
- * @returns {JSX.Element} Rendered minidenticon image.
- */
-
-export const MinidenticonImg = React.memo(({ username, saturation, lightness, ...props}) => {
-  const svgURI = useMemo(
-    () => 'data:image/svg+xml;utf8,' + encodeURIComponent(minidenticon(username, saturation, lightness)),
-    [username, saturation, lightness]
-  )
-
-  // console.log('SVG URI:', svgURI) // test for rerenders of variable
-  return (<img src={svgURI} alt={username} {...props} />)
-});
-
-/**
- * Handles client information state and other interactions with the landing page.
- * @returns {JSX.Element} Rendered App component.
- */
+import Welcome from './Welcome.jsx';
 
 const App = () => {
   const [user, setUser] = useState('');
   const [userColor, setUserColor] = useState('')
   const [lobby, setLobby] = useState('');
-  const [showLobby, setShowLobby] = useState(false);
-  const socket = useRef(null);
   const [muted, setMuted] = useState(false);
-  const [playEnter] = useSound(Enter, {volume: muted ? 0: 0.2});
-  const [playClick] = useSound(Click, {volume: muted ? 0: 0.2});
-  const [playDenied] = useSound(Denied, {volume: muted ? 0: 0.2});
-  const maxLength = 16;
-  
+  // socket data needs to be accessible by other components through socket.current; Lobby.jsx after the call to join a lobby within Welcome.jsx
+  const socket = useRef(null);
+
   const connectWebSocket = () => {
     // need to wait until connection is completed. async/await syntax not supported by WebSockets
     return new Promise((resolve, reject) => {
@@ -70,139 +41,66 @@ const App = () => {
       socket.current = new WebSocket(wsPath);
 
       socket.current.onopen = (e) => {
-        console.log('WebSocket connected');
         resolve();
+        console.log('WebSocket connected');
+        // send lobby information to the server
+        socket.current.send(JSON.stringify({action: "join", user, lobby}));
       };
 
       socket.current.onclose = (e) => {
-        console.log('WebSocket closed');
+        console.log('WebSocket closed code: ', e.code)
         reject(new Error('WebSocket closed'));
       };
     });
   };
 
-  const joinLobby = async (e) => {
-    if(user !== "" && lobby !== "") {
-      if(user.length > 16 || lobby.length > 16) {
-        console.error('Username and lobby name should be 16 characters or less.')
-        return;
-      }
-      try {
-        console.log('Attempting to join lobby...')
-        // set user color for the lobby
-        const generatedAvatar = minidenticon(user, '90', '55')
-        const match = /fill="([^"]+)"/.exec(generatedAvatar);
-        const extractedColor = match ? match[1] : 'defaultColor';
-        setUserColor(extractedColor);
-        
-        // connect WebSocket when the user joins a lobby
-        await connectWebSocket();
-        // send lobby information to the server
-        socket.current.send(JSON.stringify({action: "join", user, lobby}));
-        playEnter();
-        // then switch display to lobby
-        setShowLobby(true);
-      } catch (error) {
-        console.error('Error joining lobby:', error.message);
-      }
-    } else {
-      playDenied();
-    }
-  };
-
-  const toggleMute = () => {
-    setMuted((prevMuted) => !prevMuted);
-    // stop sounds whether muting or unmuting
-    stop();
-    if(muted) {
-      playClick();
-    }
-  }
-
   useEffect(() => {
-    // clean up WebSocket connection when the component unmounts or leaving the lobby
+    // clean up WebSocket connection when the application unmounts
     return () => {
+      // need socket closure if user returns to the landing page from a lobby
       if(socket.current) {
         console.log('Closing WebSocket connection...')
         socket.current.close();
       }
+      //otherwise, socket closure is handled server-side upon losing connection to client
     };
   }, []);
 
-  // 'Enter' should only be used for trying to enter a lobby
-  useEffect(() => {
-    const handleKeyPress = (e) => {
-      if(e && e.key === 'Enter') {
-        e.preventDefault();
-        joinLobby();
-      }
-    };
-
-    document.addEventListener('keypress', handleKeyPress);
-
-    return () => {
-      document.removeEventListener('keypress', handleKeyPress);
-    };
-  }, [joinLobby])
-
   return (
-    <div className='App'>
-      {showLobby ? ( // show lobby when userID is received
+    <Routes>
+      <Route
+        exact path="/"
+        element={
+          <Welcome
+            connectWebSocket={connectWebSocket}
+            socket={socket.current}
+            user={user}
+            setUserColor={setUserColor}
+            lobby={lobby}
+            setLobby={setLobby}
+            setUser={setUser}
+            muted={muted}
+            setMuted={setMuted}
+          />
+        }
+      />
+      <Route
+        path="/lobby"
+        element={
           <Lobby
-            socket={socket}
+            socket={socket.current}
             user={user}
             userColor={userColor}
             lobby={lobby}
-            setUser={setUser}
             setLobby={setLobby}
-            setShowLobby={setShowLobby}
+            setUser={setUser}
+            muted={muted}
+            setMuted={setMuted}
           />
-        ) : (
-          <div className='welcome-container'>
-            <h1 className='app-h'>Word Roulette Go!</h1>
-            <h3 className='app-sh'>create or join a lobby.</h3>
-            <div className='app-input'>
-              <div className='app-user'>
-                <p className='no-select'>Username:</p>
-                <textarea
-                  className='app-textarea'
-                  value={user}
-                  onChange={(e) => setUser(e.target.value)}
-                  placeholder='Choose your username!'
-                  maxLength={maxLength}
-                />
-              </div>
-              <div className='app-lobby'>
-                <p className='no-select'>Lobby Name:</p>
-                <textarea
-                  className='app-textarea'
-                  value={lobby}
-                  onChange={(e) => setLobby(e.target.value)}
-                  placeholder='Create or join a lobby!'
-                  maxLength={maxLength}
-                />
-              </div>
-              <div className='app-enter-container'>
-                <MinidenticonImg
-                  style={{ visibility: user !== '' ? 'visible' : 'hidden'}}
-                  className="app-avatar"
-                  username={user}
-                  saturation="90"
-                  lightness="55"
-                />
-                <button className='app-enter' onClick={joinLobby}>ENTER</button>
-                <button className='toggle-mute' onClick={toggleMute}>
-                  <img
-                    src={muted ? mutedSVG : unmutedSVG}
-                    alt={muted ? 'Unmute' : 'Mute'}
-                  />
-                </button>
-              </div>
-            </div>
-          </div>
-        )}
-      </div>
-    )
-  };
+        }
+      />
+    </Routes>
+  );
+};
 
 export default App;
