--- frontend/src/styles/app.scss
@@ -1,130 +1,0 @@
-/* styling for App.js */
-.welcome-container {
-  display: grid;
-  grid-template-rows: repeat(3, auto);
-  justify-content: center;
-  align-items: center;
-}
-
-.app-h {
-  display: grid;
-  justify-content: center;
-  color: rgb(226, 245, 218);
-  text-shadow: 0px 2px 4px rgb(56, 56, 56);
-  font-family: "GohuFont uni14 Nerd Font Mono";
-  font-size: 55px;
-  margin-top: 20vh;
-  margin-bottom: 0px;
-  user-select: none;
-}
-
-.app-sh {
-  display: flex;
-  justify-content: center;
-  color: rgb(218, 240, 245);
-  text-shadow: 0px 2px 4px rgb(56, 56, 56);
-  font-family: "GohuFont uni14 Nerd Font Mono";
-  font-size: 40px;
-  margin-top: 15px;
-  user-select: none;
-}
-
-.toggle-mute {
-  border: none;
-  background: none;
-  cursor: pointer;
-  user-select: none;
-}
-
-.app-input {
-  display: grid;
-  grid-template-rows: repeat(3, auto);
-  grid-template-columns: 400px;
-  justify-content: center;
-  align-items: center;
-  font-family: "ProggyVector";
-}
-
-.app-user,
-.app-lobby {
-  display: grid;
-  border: none;
-  outline: none;
-  align-items: center;
-  grid-template-columns: 140px 1fr;
-  color: rgb(226, 245, 218);
-  font-family: "DinaRemaster";
-  font-size: 25px;
-  gap: 20px;
-}
-
-.app-textarea {
-  height: 40px;
-}
-
-.app-enter-container {
-  display: flex;
-  justify-content: space-around;
-  user-select: none;
-}
-
-.app-enter {
-  display: grid;
-  padding-left: 15px;
-  padding-right: 15px;
-  border: none;
-  border-radius: 6px;
-  width: fit-content;
-  background-color: rgb(226, 245, 218, 0.775);
-  color: rgb(34, 35, 33);
-  text-shadow: 0px 2px 4px rgb(218, 240, 245, 0.775);
-  font-family: "GohuFont uni14 Nerd Font Mono";
-  font-weight: bold;
-  font-size: 25px;
-  place-self: center;
-  cursor: pointer;
-  /* margin-top: 15px; */
-  transition: 0.2s;
-}
-
-.app-enter:hover {
-  transform: scale(1.05);
-}
-
-/* small screen/mobile alterations */
-@media only screen and (max-width: 668px) {
-  .welcome-container {
-    display: grid;
-    grid-template-rows: repeat(2, 110px);
-    justify-content: center;
-    align-items: center;
-    height: 95vh;
-  }
-
-  .app-h {
-    display: grid;
-    justify-content: center;
-    color: rgb(226, 245, 218);
-    text-shadow: 0px 2px 4px rgb(56, 56, 56);
-    font-family: "GohuFont uni14 Nerd Font Mono";
-    font-size: 34px;
-  }
-
-  .app-sh {
-    display: flex;
-    justify-content: center;
-    color: rgb(218, 240, 245);
-    text-shadow: 0px 2px 4px rgb(56, 56, 56);
-    font-family: "GohuFont uni14 Nerd Font Mono";
-    font-size: 22px;
-  }
-
-  .app-input {
-    display: grid;
-    grid-template-rows: repeat(3, auto);
-    grid-template-columns: 350px;
-    justify-content: center;
-    align-items: center;
-    font-family: "ProggyVector";
-  }
-}
