+++ frontend/src/styles/global.scss
@@ -95,11 +95,11 @@ a {
   bottom: 0;
   z-index: 1;
     
-  -moz-animation:move-background 70s linear infinite;
-  -ms-animation:move-background 70s linear infinite;
-  -o-animation:move-background 70s linear infinite;
-  -webkit-animation:move-background 70s linear infinite;
-  animation:move-background 70s linear infinite;
+  -moz-animation:move-background 60s linear infinite;
+  -ms-animation:move-background 60s linear infinite;
+  -o-animation:move-background 60s linear infinite;
+  -webkit-animation:move-background 60s linear infinite;
+  animation:move-background 60s linear infinite;
 }
 
 .no-select {
