+++ todo.md
@@ -6,14 +6,16 @@
 - [x] Increase size of input box when typing a message in lobby
 - [x] Limit max username/lobby name length (16 char at the moment)
 - [ ] Limit max number of messages in the chatroom
-- [ ] Add onto an already-sent message if the immediate next message is sent by the same user (prevents unnecessary repetition of message-info)
+- [x] Add onto an already-sent message if the immediate next message is sent by the same user (prevents unnecessary repetition of message-info)
 
 ## FIX
 - [x] usage of dotenv for external ip in the socket connection from the frontend (possibly find solution not dotenv?)
 - [ ] mobile UI (honestly the UI in general as well lol)
 - [ ] pay attention to rerendering amount intermittently (keep resources low)
-- [ ] possibly render user text immediately instead of waiting for sync with db (probably bad practice though)
+<!-- - [ ] possibly render user text immediately instead of waiting for sync with db (probably bad practice though) -->
+- [ ] Holding 'Shift' + 'Enter' when sending a message was sending unknown actions to the backend. Not allowing users to insert new lines into their messages as a temporary solution. But want to support normal functionality at some point.
 
 ## REFACTOR
 - [x] A WebSocket connection should only be instantiated when a user enters a lobby, not upon coming to the site itself
 - [x] Reduce served font file sizes
+- [ ] Use React Router as good practice instead of conditional.
